Initial purge of QTKit and movie support.
Not fully complete yet. Ideally we can replace it with AVFoundation,
but that will require a bigger refactor. For now, SFBAudioEngine may
be adequate.

Also removed is movie support. The support for it seemed planned, but
never actually done, so the nibs/code for it was basically all nopped
out. Let's rip it out and revisit it later - I won't use it, since
Navidrome doesn't do movies.