Purge the hot key handlers
These are super grody, and nowadays with the Now Playing system wide,
I don't think it's needed. If it turns out to be desired, it probably
needs a serious rewrite anyways, considering these likely go raw on
APIs that are now verboten accept for a11y, which is something I'm not
comfortable with. Worst case, I could ask Chris et al what they do for
Cog.

I have not touched the NIBs yet because I don't want to do any Xcode
specific stuff just yet.

I also plan to get rid of the menu bar icon, because that's another
thing that can be replaced with the Now Playing thing.