Menu bar stuff for switching views
These aren't terribly common operations IMHO, so not having in toolbar
or source list is probably tolerable. Could think of a better place for
it...

Note that the search view is not accessible this way, but instead is
invoked via the search bar. It could do with a menu item though...

User list I plan to put into the sidebar.