Pause playback progress timer when not in use
Should improve battery life.

Ideally, we should be using observers, but this is complex with two
possible implementations for our backend, even if we aren't using
SFBAudioPlayer anymore.

The "now playing" update timer also happens still, but it happens at
a much less active rate, so the impact is less.

Fixes GH-12