Show per-track artist info
It seems that the artist info for a track is in artistName, and the
album artist is the SBArtist associated to the SBAlbum. For the server
library view, show the artist of each track in the table, since it's
otherwise redundant.

Local music for now has a nil artistName, so it'll require a refactor
of import for new tracks to consider album SBArtist vs. artistName
string - and this could also be an SBArtist in the future. We could
also consider a flag day, or use a transformer to fall back better.

Relates to GH-34