Create monospace text class, use it in nib
Annoyingly, we can't specify the font of this in IB. Instead, we have
to do it programmatically. It seems setting `font` is unreliable, so
we have to make an attributed string. (If you set font, it seems to
get aggressively set back.)

This uses Swift, which is inevitable, so let's try it with something
easy. I suspect more Swift will come in as needed.

Fixed GH-21