Experiments with using AudioToolbox for metadata
Instead of SFBAudioFile. This isn't hooked up yet, just printing to log,
and is also pretty nasty. I'll encapsulate this better, and it'll allow
for nicer strategies to be employed. (That, or it can be ripped out and
using TagLib, but AT and maybe AVAsset gets us most of the way there for
something I suspect will be rarely used.)