Workaround for album without parents showing up
They come from getAlbumList - an example is the demo server returning
Movies, which is in the root dir. This means it doesn't have a parent,
so it'll be nil. This is really confusing.

It does mean we keep enforcing our assumptions of directory structure
for Subsonic (but not Navidrome) users, annoyingly. We should /really/
move to using the ID3 endpoints.