mitigate issue with SplitView autosaving for wrong dimensions
AppKit will try to fit you in the safe area if you have a full frame,
but NSSplitView doesn't know about this. We have to give it a little
help.

Local music doesn't autosave its position, so it's immune. The server
album view does, but we have to force its frame into the safe area,
so it won't trigger this issue. However, if local music does autosave,
it'll need this fix too, so it can be refactored out.