Fix up table views
Totally embarassing: There were duplicate identifiers in the nibs.
That really confuses the hell out of AppKit when persisting columns.
I'm surprised that it didn't log out to the console for this though.

In addition to that, remove unneeded code in SBTableView since macOS
can do the width/visibility persistence itself at least in 13.

Also fix the sizing of headers in the nibs too.

Fixes GH-146