Use index based tracklist
This is motivated by the fact the tracklist was based on the current
track. However, if there were duplicate tracks, then it was confusing as
to what track was playing and playing next.

This complicates some stuff in SBPlayer to use an index based approach,
with the benefit of making things less confusing for users.