Actually unbreak podcast UI
This was never working in the first place. The podcast menu item didn't
switch, and items were duplicating. And the stream URL was for the wrong
ID type. Things now work, so it's at least better than it used to be.

Still some grungy stuff - there's no way to manipulate the podcasts, for
instance. And the automatic download stuff has some curious ideas about
metadata? Lots of UI thoughts for anywhere where podcasts intersect w/
noormal tracks.

(Or maybe we can remove it?)