Don't show an alert when removing missing items
When transitioning to the new schema, you'll likely get albums that
don't exist, though we try to update as soon as possible when the artist
album list is pulled. My guess is a race condition between getting the
album list from the server and the cached albums from Core Data.

The easiest workaround is to just not show this message.