Simplify split view logic
Use holding priorities when possible, introduced in 10.8. These should
clamp the sizes properly, except the special case of the library view,
where otherwise, the split view would resize on first becoming visible
or when the right sidebar was toggled in or out of view. To fix this, we
implement the delegate method of shouldAdjustSizeOfSubview:, which while
not recommended for auto layout views, does work, and prevents the
unexpected sizing while keeping auto layout working.