Finagle the sort order a little
Navidrome at least has an [Unknown] index with an [Unknown Artist]
artist entry below it, but the Foundation sorting puts it in a clumsy
place with the index and artist flipped. Provide a custom sort order
that deals with it. Ugly and Navidrome specific, but does work.

I'm considering preserving the order the server gives us, but that has
its own problems (is it actually desirable?). Might need to rethink the
schema with index vs. artist a little.