Move a bunch of common behaviours into the superclasses
Since the superclass implements them and is now aware of the selection,
this means we can reduce a ton of duplicate (and subtly different!)
code. Some stuff that requires VC-specific behaviour (like show
albums/artists in finder for local music, deletion for
local/playlist/tracklist, etc.) can be overriden.

Some stuff has not been moved yet - the playback requires knowing the
non-selected tracks, and validation can be done in the superclass (with
overrides) as well.