Beginning of supporting top tracks
This turns the search results controller into a generic "we have a list of
tracks" controller. It should probably be renamed, along with the class
and enum that holds the tracks to begin with.

There's some hacks to make this work with an ObjC frontend. Once we
convert it to Swift, we can just expose the enum directly.

UI is not wired up yet.