Fix covers getting deleted in macOS 15
UserNotifications now deletes the attachments on macOS 15 (or did
before, and instead of us never triggering, does so automatically).

Regardless, this is not good for us. Instead of using the image on disk
we use for everything else, make a copy instead in the temporary
directory. On >=15, it'll delete that, and before that, it should get
cleaned automatically anyways.