Improve cover path handling
I've inherited a dog's breakfast and kept making the dog breakfast. We
shouldn't have absolute paths in the database, and we shouldn't have the
get code do setting to do so.

This adds an operation that runs on launch to clean up the database
instead of doing it on demand (which could also cause some confusion).
It also handles cross-linked files (i.e. local library item referencing
cover in the subdirectory for another server), and things like incorrect
absolute paths.

It may be worth setting the path regardless, or to nil if we can't copy
the item.