Add view to show all tracks on a server
This takes advantage of the fact OpenSubsonic specifies that an
empty query should return all tracks. With infinite scrolling, this
means we can just reuse the search controller and have tracks page in as
needed without having a massive download.

This also makes it easy to i.e. download all tracks (fixes GH-147).