Finally have working editing
I still don't really understand how view based table editing is supposed
to work. This also rewrites it to avoid SwiftUI, since the responder
chain games we had to play made it look ugly AF. SwiftUI was actually
pretty simple for layout compared to constraints, so I'm tempted to go
back to this.

This also moves the rename logic too.