Add more properties
Mostly OpenSubsonic properties that don't require custom parsing or much
thought into schemas. Also shows them in the inspector.

These are *not* scalar, because those are i.e. Int, not Int?, and
optional means that it defaults to 0. Not ideal.

Artist/album properties are not shown yet. That may take adding the
description for artist/album info, which, the changes in parsing for
artists should makee that easier.