Fix VoiceOver not reading the albums in collection
It turns out this was a little bit probably missing the accessibility
label since we roll our own, but mostly because of the ugly workaround
we were using. It turns out that you can't actually have the hosting
view be the view for the view item, but rather as a subview. This allows
us to use the normal reuse lifecycle, and with that, it seems
accessibility works correctly out of the box.

This also adds a little bit of a11y label in case SwiftUI is a bit
stinky with automatic labels. It seems to work correctly in VoiceOver.

There may be other places this can be improved.

Fixes GH-239.